Move build directory
