+++ package.json
@@ -1,12 +1,12 @@
 {
   "name": "@emcasa/mock-server",
   "version": "0.1.0",
-  "main": "build/index.js",
+  "main": "lib/index.js",
   "author": "EmCasa <dev@emcasa.com>",
   "license": "MIT",
   "scripts": {
     "start": "nodemon src/run.js",
-    "build": "babel -d build src",
+    "build": "babel -d lib src",
     "update-schema": "cd backend && mix absinthe.schema.json --schema ReWeb.Schema --pretty && mv schema.json ../schema.json",
     "prepare-backend": "git submodule init && (cd backend && cp -n config/dev.secret-example.exs config/dev.secret.exs || mix deps.get)",
     "prepare": "yarn prepare-backend && yarn build"
