+++ src/graphql/mocks/Address.js
@@ -1,7 +1,7 @@
 import flow from "lodash.flow";
 import faker from "faker/locale/pt_BR";
-import slugify from "../helpers/slugify";
-import withSeed from "../helpers/withSeed";
+import slugify from "../../helpers/slugify";
+import withSeed from "../../helpers/withSeed";
 import { DISTRICTS } from "./District";
 
 const neighborhood = () => faker.random.arrayElement(DISTRICTS);
