+++ src/graphql/mocks/User.js
@@ -1,7 +1,7 @@
-export default (_, __, { user = {} } = {}) => ({
+export default (_, __, { user } = {}) => ({
   name: "John Doe",
   id: 300,
   email: "johndoe@example.com",
   phone: "+552199999999",
-  ...user
+  ...(user || {})
 });
