+++ src/index.js
@@ -1,7 +1,1 @@
 export { default as createApolloServer } from "./graphql";
-export {
-  encodeJwt,
-  decodeJwt,
-  encodeAccessToken,
-  decodeAccessToken
-} from "./graphql/helpers/jwt";
